+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.1.0
 =====
 * First version of plugin architecture
+* Make Jenkins rules optional
 
 v0.0.1
 =====